frontend/requirements.txt
streamlit==1.28.2
python-dotenv==1.0.0
openai>=1.0.0
